+++ include/chemfiles/formats/Molfile.hpp
@@ -61,8 +61,8 @@ public:
     ~Molfile() override;
 
     void read(Frame& frame) override;
-    void read_step(size_t step, Frame& frame) override;
-    size_t nsteps() override;
+    void read_at(size_t index, Frame& frame) override;
+    size_t size() override;
 private:
     /// Convert a molfile timestep to a chemfiles frame
     void molfile_to_frame(const molfile_timestep_t& timestep, Frame& frame);