+++ include/chemfiles/formats/TPR.hpp
@@ -28,9 +28,9 @@ class TPRFormat final : public Format {
   public:
     TPRFormat(std::string path, File::Mode mode, File::Compression compression);
 
-    void read_step(size_t step, Frame& frame) override;
+    void read_at(size_t index, Frame& frame) override;
     void read(Frame& frame) override;
-    size_t nsteps() override;
+    size_t size() override;
 
   private:
     // Since GROMACS 2020 (TPR version 119) the way the body is deserialized changes.
@@ -126,8 +126,8 @@ class TPRFormat final : public Format {
     XDRFile file_;
     /// TPR header of the file
     TprHeader header_;
-    /// The next step to read
-    size_t step_ = 0;
+    /// index of the next step to read
+    size_t index_ = 0;
 };
 
 template <> const FormatMetadata& format_metadata<TPRFormat>();