+++ src/formats/TNG.cpp
@@ -91,9 +91,9 @@ TNGFormat::TNGFormat(std::string path, File::Mode mode, File::Compression compre
 
     // work around a bug in tng_num_frames_get (https://redmine.gromacs.org/issues/2937)
     // manually query all frames in the trajectory, and get the corresponding
-    // TNG frame number in tng_steps_
-    int64_t current_frame = -1;
-    int64_t next_frame = 0;
+    // TNG frame number in simulation_steps_
+    int64_t current_step = -1;
+    int64_t next_step = 0;
     int64_t unused = 0;
     int64_t* buffer = nullptr;
 
@@ -102,12 +102,12 @@ TNGFormat::TNGFormat(std::string path, File::Mode mode, File::Compression compre
         // Look for all frames with at least position data
         int64_t block_ids[] = {TNG_TRAJ_POSITIONS};
         status = tng_util_trajectory_next_frame_present_data_blocks_find(
-            tng_, current_frame, 1, block_ids, &next_frame, &unused, &buffer
+            tng_, current_step, 1, block_ids, &next_step, &unused, &buffer
         );
 
         if (status == TNG_SUCCESS) {
-            current_frame = next_frame;
-            tng_steps_.push_back(current_frame);
+            current_step = next_step;
+            simulation_steps_.push_back(current_step);
         } else if (status == TNG_FAILURE) {
             // We found the end of the file
             break;
@@ -120,17 +120,17 @@ TNGFormat::TNGFormat(std::string path, File::Mode mode, File::Compression compre
     free(buffer);
 }
 
-size_t TNGFormat::nsteps() {
-    return tng_steps_.size();
+size_t TNGFormat::size() {
+    return simulation_steps_.size();
 }
 
-void TNGFormat::read_step(size_t step, Frame& frame) {
-    step_ = step;
+void TNGFormat::read_at(size_t index, Frame& frame) {
+    index_ = index;
     read(frame);
 }
 
 void TNGFormat::read(Frame& frame) {
-    frame.set("simulation_step", tng_steps_[step_]);
+    frame.set("simulation_step", simulation_steps_[index_]);
     natoms_ = 0;
     CHECK(tng_num_particles_get(tng_, &natoms_));
     assert(natoms_ > 0);
@@ -138,7 +138,7 @@ void TNGFormat::read(Frame& frame) {
 
     double time = 0;
     tng_function_status status =
-        tng_util_time_of_frame_get(tng_, tng_steps_[step_], &time);
+        tng_util_time_of_frame_get(tng_, simulation_steps_[index_], &time);
     if (status == TNG_SUCCESS) {
         // TNG stores time in seconds
         // convert to pico seconds
@@ -150,7 +150,7 @@ void TNGFormat::read(Frame& frame) {
     read_cell(frame);
     read_topology(frame);
 
-    step_++;
+    index_++;
 }
 
 void TNGFormat::read_positions(Frame& frame) {
@@ -158,7 +158,7 @@ void TNGFormat::read_positions(Frame& frame) {
     int64_t unused = 0;
 
     CHECK(tng_util_pos_read_range(
-        tng_, tng_steps_[step_], tng_steps_[step_], buffer.ptr(), &unused
+        tng_, simulation_steps_[index_], simulation_steps_[index_], buffer.ptr(), &unused
     ));
 
     auto positions = frame.positions();
@@ -174,7 +174,7 @@ void TNGFormat::read_velocities(Frame& frame) {
     int64_t unused = 0;
 
     auto status = tng_util_vel_read_range(
-        tng_, tng_steps_[step_], tng_steps_[step_], buffer.ptr(), &unused
+        tng_, simulation_steps_[index_], simulation_steps_[index_], buffer.ptr(), &unused
     );
 
     switch (status) {
@@ -208,7 +208,7 @@ void TNGFormat::read_cell(Frame& frame) {
     int64_t unused = 0;
 
     auto status = tng_util_box_shape_read_range(
-        tng_, tng_steps_[step_], tng_steps_[step_], buffer.ptr(), &unused
+        tng_, simulation_steps_[index_], simulation_steps_[index_], buffer.ptr(), &unused
     );
 
     switch (status) {