+++ tests/doc/trajectory/size.cpp
@@ -10,13 +10,10 @@ TEST_CASE() {
     // [example]
     auto trajectory = Trajectory("water.nc");
 
-    auto frame = trajectory.read_step(4);
-    // Use the frame for awesome science here!
-
-    // This is one way to iterate over all the frames in a trajectory
-    for (size_t i = 0; i < trajectory.nsteps(); i++) {
-        frame = trajectory.read_step(i);
-        // ...
+    auto size = trajectory.size();
+    for (size_t i = 0; i < size; i++) {
+        auto frame = trajectory.read_at(i);
     }
+
     // [example]
 }